requirements-dev.txt
# Requirements needed for running the test suite.

coverage==3.7.1
django-nose==1.4
factory_boy==2.5.2
flake8==2.4.1
mccabe==0.2.1
mock==1.0.1
nose==1.3.6
pep8==1.5.7
pyflakes==0.8.1
six==1.9.0
